requirements.txt
dnspython==2.1.0
google-api-python-client==2.31.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
oauth2client==4.1.3
pandas==1.2.1
pre-commit==2.13.0
pydantic==1.8.2
python-dotenv==0.17.1
python-json-logger==2.0.1
pymongo==3.12
